requirements.txt
streamlit
pandas
matplotlib
altair
pillow
opencv-python-headless
numpy
torch
torchvision
grad-cam